AdvantagePool, various fixes
- fixed issue with graph of "at most" and "at least" PMFs
- clarified various variable names
- greatly improved documentation for DicePool.__pool_convolution__()
- trimmed unnecessary conditional branch in DicePool.__pool_convolution__()
- implemented AdvantagePool() class
- updated readme

AdvantagePool implements functionality for rolling a pool of N dice of size S and then keeping the highest (or lowest) result of the pool. Solution implemented algebraically.

Functionality for keeping the highest (or lowest) K results of the pool and summing them together into a final result has not yet been implemented. Awaiting an analytical solution.